Move alerts and pulse components to notifications domain
